requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
openai
pydantic
python-dotenv
